+++ compile/detectTransactionExampleNumbers.js
