+++ index.js
@@ -1,0 +1,5 @@
+const parse = require('./parse');
+const compile = require('./compile');
+
+
+module.exports = { parse, compile };
