+++ scripts/smoke.sh
@@ -56,10 +56,29 @@ if [ ! -z "$TRAVIS" ]; then
       "The packaging of Dredd Transactions should have prevented this."
     echo "======================================================================"
     exit 1
-  else
-    echo "SUCCESS"
-    echo "======================================================================"
   fi
+
+  echo "Importing dredd-transactions"
+  echo "======================================================================"
+  echo "const assert = require('assert');" > index.js
+  echo "const dt = require('dredd-transactions');" >> index.js
+  echo "assert.ok(typeof dt.parse === 'function');" >> index.js
+  echo "assert.ok(typeof dt.compile === 'function');" >> index.js
+  node index.js
+
+  echo "======================================================================"
+  echo "Importing dredd-transactions/parse and dredd-transactions/compile"
+  echo "======================================================================"
+  echo "const assert = require('assert');" > index.js
+  echo "const parse = require('dredd-transactions/parse');" >> index.js
+  echo "const compile = require('dredd-transactions/compile');" >> index.js
+  echo "assert.ok(typeof parse === 'function');" >> index.js
+  echo "assert.ok(typeof compile === 'function');" >> index.js
+  node index.js
+
+  echo "======================================================================"
+  echo "SUCCESS"
+  echo "======================================================================"
 else
   exit 1
 fi
